[Update] QueryBuilderHandler component
